requirements.txt
google-auth>=2.28.0
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
PyYAML>=6.0.1
pytest>=8.0.0
